Refactor standalone e2e and add support for Windows runner (#1036)
