+++ tests/e2e/common/common.go
@@ -90,7 +90,7 @@ func GetVersionsFromEnv(t *testing.T) (string, string) {
 
 func UpgradeTest(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		args := []string{
 			"upgrade", "-k",
 			"--runtime-version", details.RuntimeVersion,
@@ -117,7 +117,7 @@ func UpgradeTest(details VersionDetails, opts TestOptions) func(t *testing.T) {
 }
 
 func EnsureUninstall(all bool) (string, error) {
-	daprPath := getDaprPath()
+	daprPath := GetDaprPath()
 
 	var _command [10]string
 	command := append(_command[0:], "uninstall", "-k")
@@ -187,7 +187,7 @@ func GetTestsPostCertificateRenewal(details VersionDetails, opts TestOptions) []
 
 func MTLSTestOnInstallUpgrade(opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "mtls", "-k")
 		require.NoError(t, err, "expected no error on querying for mtls")
 		if !opts.MTLSEnabled {
@@ -233,7 +233,7 @@ func MTLSTestOnInstallUpgrade(opts TestOptions) func(t *testing.T) {
 
 func ComponentsTestOnInstallUpgrade(opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		// if dapr is installed.
 		if opts.ApplyComponentChanges {
 			// apply any changes to the component.
@@ -256,7 +256,7 @@ func ComponentsTestOnInstallUpgrade(opts TestOptions) func(t *testing.T) {
 
 func StatusTestOnInstallUpgrade(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "status", "-k")
 		require.NoError(t, err, "status check failed")
 		var notFound map[string][]string
@@ -419,7 +419,7 @@ func CRDTest(details VersionDetails, opts TestOptions) func(t *testing.T) {
 
 func GenerateNewCertAndRenew(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		err := exportCurrentCertificate(daprPath)
 		require.NoError(t, err, "expected no error on certificate exporting")
 
@@ -445,7 +445,7 @@ func GenerateNewCertAndRenew(details VersionDetails, opts TestOptions) func(t *t
 
 func UseProvidedPrivateKeyAndRenewCerts(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		args := []string{
 			"mtls", "renew-certificate", "-k",
 			"--private-key", "../testdata/example-root.key",
@@ -473,7 +473,7 @@ func UseProvidedPrivateKeyAndRenewCerts(details VersionDetails, opts TestOptions
 
 func UseProvidedNewCertAndRenew(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		args := []string{
 			"mtls", "renew-certificate", "-k",
 			"--ca-root-certificate", "./certs/ca.crt",
@@ -507,7 +507,7 @@ func UseProvidedNewCertAndRenew(details VersionDetails, opts TestOptions) func(t
 
 func NegativeScenarioForCertRenew() func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		args := []string{
 			"mtls", "renew-certificate", "-k",
 			"--ca-root-certificate", "invalid_cert_file.pem",
@@ -551,7 +551,7 @@ func NegativeScenarioForCertRenew() func(t *testing.T) {
 
 func CheckMTLSStatus(details VersionDetails, opts TestOptions, shouldWarningExist bool) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "mtls", "-k")
 		require.NoError(t, err, "expected no error on querying for mtls")
 		if !opts.MTLSEnabled {
@@ -595,7 +595,7 @@ func (v VersionDetails) constructFoundMap(res Resource) map[string]bool {
 	return foundMap
 }
 
-func getDaprPath() string {
+func GetDaprPath() string {
 	distDir := fmt.Sprintf("%s_%s", runtime.GOOS, runtime.GOARCH)
 
 	return filepath.Join("..", "..", "..", "dist", distDir, "release", "dapr")
@@ -642,7 +642,7 @@ func getClient() (*k8s.Clientset, error) {
 
 func installTest(details VersionDetails, opts TestOptions) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		args := []string{
 			"init", "-k",
 			"--wait",
@@ -693,7 +693,7 @@ func kubernetesTestOnUninstall() func(t *testing.T) {
 	return func(t *testing.T) {
 		_, err := EnsureUninstall(true)
 		require.NoError(t, err, "uninstall failed")
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "uninstall", "-k")
 		require.NoError(t, err, "expected no error on uninstall without install")
 		require.Contains(t, output, "Removing Dapr from your cluster...", "expected output to contain message")
@@ -704,7 +704,7 @@ func kubernetesTestOnUninstall() func(t *testing.T) {
 
 func uninstallMTLSTest() func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "mtls", "-k")
 		require.Error(t, err, "expected error to be return if dapr not installed")
 		require.Contains(t, output, "error checking mTLS: system configuration not found", "expected output to match")
@@ -713,7 +713,7 @@ func uninstallMTLSTest() func(t *testing.T) {
 
 func componentsTestOnUninstall(all bool) func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		// On Dapr uninstall CRDs are not removed, consequently the components will not be removed.
 		// TODO Related to https://github.com/dapr/cli/issues/656.
 		// For now the components remain.
@@ -747,7 +747,7 @@ func componentsTestOnUninstall(all bool) func(t *testing.T) {
 
 func statusTestOnUninstall() func(t *testing.T) {
 	return func(t *testing.T) {
-		daprPath := getDaprPath()
+		daprPath := GetDaprPath()
 		output, err := spawn.Command(daprPath, "status", "-k")
 		t.Log("checking status fails as expected")
 		require.Error(t, err, "status check did not fail as expected")
